IdlePlanetMinerTools

Documentation for IdlePlanetMinerTools.

This package provides some tools for thise who play the TechTree game Idle Planet Miner.

IdlePlanetMinerTools

Things and Inventory

Recipies

Arithmetic

Projects

Modifiers

Index

Definitions

Things and Inventory

Mining, smelting and crafting produce instances of subtypes of Ore, Alloy, or Crafted, respectively. These are all subtypes of Thing. Each concrete subtype of Thing has a count field indicating a quantity of that Thing.

One can make a Thing using a concrete constructor

using IdlePlanetMinerTools

Battery(3)

Battery(3.0)

or using the @t_str macro, which makes the named thing with a count of 1.

t"battery"

Battery(1.0)

Note that for convenience, @t_str uses a dictionary of all concrate Thing subtypes and finds the closest string match.

An Inventory represents a collection of disperate Things.

For convenience, there is a total ordering defined on all concrete subtypes of Thing.

Recipies

A Recipie describes how to smelt an [Ore]((@ref) to an [Alloy](@ref), anAlloyto a [Crafted](@ref) item or to create a [Project`](@ref) from such things.

You can lookup a recipie using an rx string:

using IdlePlanetMinerTools

rx"LaserTorch"

rx"Advanced Crafter"

Recipie(AdvancedCrafter, Inventory(Thing[GoldBar(50.0), Lens(5.0)]), 0)

Arithmetic

Arithmetic operations are defined for Things and Inventorys:

using IdlePlanetMinerTools

need = 7 * delta(rx"gravity chamber")

have = GravityChamber(4) + AdvancedComputer(42) + BasicComputer(64) + Circuit(902)

actions, result = crafting_plan(have - need)

(CraftingAction[CraftingAction(Recipie(GravityChamber, Inventory(Thing[AdvancedComputer(-60.0), NuclearReactor(-1.0)]), 24300), 3)], Inventory(Thing[Circuit(902.0), BasicComputer(64.0), AdvancedComputer(282.0), NuclearReactor(4.0), GravityChamber(0.0)]))

Projects

Once you reasearch a Project the game improves on your capabilities or gives you access to additional functionality. Each Project has a Recipie that describes what you need to be able to reasearch it. Most Projects also have prerequisites which are other projects that you need to have reasearched beforehand.

Once you've researched a project, you should add it to DEFAULT_MODIFIERS so that its benefits will be used throughout this package.

Modifiers

During the course of the game, the player may accumulate status that alters calculations. Among these are Projects, Managers and space ship Rooms. Each of these is a kind of Modifier. DEFAULT_MODIFIERS is a list of modifiers that the player has said they already have. It is used by delta to determine what is needed to complete a Recipie if another list of Modifiers is not passed to it.

Index

IdlePlanetMinerTools.DEFAULT_MODIFIERS

IdlePlanetMinerTools.Alloy

IdlePlanetMinerTools.Crafted

IdlePlanetMinerTools.Inventory

IdlePlanetMinerTools.Modifier

IdlePlanetMinerTools.Ore

IdlePlanetMinerTools.Project

IdlePlanetMinerTools.Recipie

IdlePlanetMinerTools.Room

IdlePlanetMinerTools.Thing

IdlePlanetMinerTools.all_things

IdlePlanetMinerTools.base_effect

IdlePlanetMinerTools.base_selling_price

IdlePlanetMinerTools.crafting_plan

IdlePlanetMinerTools.delta

IdlePlanetMinerTools.max_level

IdlePlanetMinerTools.next_projects

IdlePlanetMinerTools.ordinal

IdlePlanetMinerTools.per_level_effect

IdlePlanetMinerTools.prerequisites

IdlePlanetMinerTools.room_effect_factor

Definitions

IdlePlanetMinerTools.DEFAULT_MODIFIERS — Constant

DEFAULT_MODIFIERS is the list of Modifiers that is used for operations like deltawhere none are specified.

IdlePlanetMinerTools.Alloy — Type

Alloy is the abstract supertype of anything that is smelted from Ore or other Alloys.

IdlePlanetMinerTools.Crafted — Type

Crafted is the abstract suppertype of anything that is crafted from other Things.

IdlePlanetMinerTools.Inventory — Type

An Inventory represents a collection of Things.

IdlePlanetMinerTools.Modifier — Type

Modifier is the abstract supertype for anything that can modify various arithmetic factors in the game, for example Projects or Rooms that modify smelting ingredients.

IdlePlanetMinerTools.Ore — Type

Ore is the abstract supertype for anything that is mined.

IdlePlanetMinerTools.Project — Type

Project represents an Idle Planet Miner project.

IdlePlanetMinerTools.Recipie — Type

Recipie describes how to make Alloys, crafted items, projects, etc, anything that required some number of Things.

IdlePlanetMinerTools.Room — Type

Room is the abstract supertype for all space ship rooms. Each room has a level field.

Associated with each Room type is a base_effect and a per_level_effect, and a max_level.

IdlePlanetMinerTools.Thing — Type

Thing is the abstract supertype of Ore, Alloy and Crafted.

Every subtype of Thing has a count field.

IdlePlanetMinerTools.all_things — Method

all_things()

returns a list of all concrete subtypes of Thing in canonical order.

IdlePlanetMinerTools.base_effect — Function

base_effect(::Type{<:Room})

Returns the modifier factor for the room type at level 1.

IdlePlanetMinerTools.base_selling_price — Method

base_selling_price(::Type{<:Thing})
base_selling_price(::Thing)

Returns the base selling price for the thing.

For instances it multiplies the unit price by the count.

IdlePlanetMinerTools.crafting_plan — Method

crafting_plan(inv::Inventory)

Appliies Recipies to the ~Inventory` until there are no negative item counts or there is no known recipie to resolve the current shortage.

Returns a list of CraftingActions and the final Inventory.

IdlePlanetMinerTools.delta — Function

delta(r::Recipie, modifiers = DEFAULT_MODIFIERS)

Returns an Inventory that would be the effect of applying the Recipie. modifiers is a vector of the player's current Modifiers.

IdlePlanetMinerTools.max_level — Function

max_level(::Type{<:Room})

Returns the maximum allowed value for the level field for the specified room type.

IdlePlanetMinerTools.next_projects — Function

next_projects(modifiers=DEFAULT_MODIFIERS)

Assuming that the projects in modifiers have already been researches, what are the next projects to be researches?

IdlePlanetMinerTools.ordinal — Method

ordinal(thing)

Returns the ordinal number for thing, where thing is either a subtype of Thing or an instance of a subtype of Thing.

IdlePlanetMinerTools.per_level_effect — Function

per_level_effect(::Type{<:Room})

Returns the modifier factor for the room type for each level above level 1.

IdlePlanetMinerTools.prerequisites — Method

prerequisites(::Type{<:Project})

Returns a vector of the immediate prerequisites that are required before researching the specified Project.

IdlePlanetMinerTools.room_effect_factor — Method

room_effect_factor(::Room)

Computes the effect factor for the Room based on its type and level. This says nothing about what is modified. It just represents a calculation that tis common to all rooms.