IdlePlanetMinerTools

Documentation for IdlePlanetMinerTools.

This package provides some tools for thise who play the TechTree game Idle Planet Miner.

IdlePlanetMinerTools

Things and Inventory

Recipies

Arithmetic

Projects

Modifiers

Index

Definitions

Things and Inventory

Mining, smelting and crafting produce instances of subtypes of Ore, Alloy, or Crafted, respectively. These are all subtypes of Thing. Each concrete subtype of Thing has a count field indicating a quantity of that Thing.

One can make a Thing using a concrete constructor

using IdlePlanetMinerTools

Battery(3)

Battery(3.0)

or using the @t_str macro, which makes the named thing with a count of 1.

t"battery"

Battery(1.0)

Note that for convenience, @t_str uses a dictionary of all concrate Thing subtypes and finds the closest string match.

An Inventory represents a collection of disperate Things.

For convenience, there is a total ordering defined on all concrete subtypes of Thing.

Recipies

A Recipie describes how to smelt an [Ore]((@ref) to an [Alloy](@ref), anAlloyto a [Crafted](@ref) item or to create a [Project`](@ref) from such things.

You can lookup a recipie using an rx string:

using IdlePlanetMinerTools

rx"LaserTorch"

rx"Advanced Crafter"

Recipie(AdvancedCrafter, Inventory(Thing[GoldBar(50.0), Lens(5.0)]), 0)

Arithmetic

Arithmetic operations are defined for Things and Inventorys:

using IdlePlanetMinerTools

need = 7 * delta(rx"gravity chamber")

have = GravityChamber(4) + AdvancedComputer(42) + BasicComputer(64) + Circuit(902)

actions, result = crafting_plan(have - need)

(CraftingAction[CraftingAction(Recipie(GravityChamber, Inventory(Thing[AdvancedComputer(-60.0), NuclearReactor(-1.0)]), 24300), 3)], Inventory(Thing[Circuit(902.0), BasicComputer(64.0), AdvancedComputer(282.0), NuclearReactor(4.0), GravityChamber(0.0)]))

Projects

Once you reasearch a Project the game improves on your capabilities or gives you access to additional functionality. Each Project has a Recipie that describes what you need to be able to reasearch it. Most Projects also have prerequisites which are other projects that you need to have reasearched beforehand.

Once you've researched a project, you should add it to DEFAULT_MODIFIERS so that its benefits will be used throughout this package.

Modifiers

During the course of the game, the player may accumulate status that alters calculations. Among these are Projects, Managers and space ship Rooms. Each of these is a kind of Modifier. DEFAULT_MODIFIERS is a list of modifiers that the player has said they already have. It is used by delta to determine what is needed to complete a Recipie if another list of Modifiers is not passed to it.

Index

IdlePlanetMinerTools.DEFAULT_MODIFIERS

IdlePlanetMinerTools.AdvancedAlchemy

IdlePlanetMinerTools.AdvancedAlloyValue

IdlePlanetMinerTools.AdvancedAnomalyResolution

IdlePlanetMinerTools.AdvancedAsteroidHarvester

IdlePlanetMinerTools.AdvancedAsteroidScanner

IdlePlanetMinerTools.AdvancedCargoHandling

IdlePlanetMinerTools.AdvancedContractManager

IdlePlanetMinerTools.AdvancedCrafter

IdlePlanetMinerTools.AdvancedDebrisHarvester

IdlePlanetMinerTools.AdvancedEchoFiltering

IdlePlanetMinerTools.AdvancedFurnace

IdlePlanetMinerTools.AdvancedFurnaceOverdrive

IdlePlanetMinerTools.AdvancedHarmonicCalibration

IdlePlanetMinerTools.AdvancedItemValue

IdlePlanetMinerTools.AdvancedLeader

IdlePlanetMinerTools.AdvancedManagerTraining

IdlePlanetMinerTools.AdvancedMarketAccelerator

IdlePlanetMinerTools.AdvancedMarketManipulation

IdlePlanetMinerTools.AdvancedMining

IdlePlanetMinerTools.AdvancedOreTargeting

IdlePlanetMinerTools.AdvancedReactiveSweep

IdlePlanetMinerTools.AdvancedRoverResupply

IdlePlanetMinerTools.AdvancedSignalStratification

IdlePlanetMinerTools.AdvancedSpecialistUniversity

IdlePlanetMinerTools.AdvancedSurgeMapping

IdlePlanetMinerTools.AdvancedTerraforming

IdlePlanetMinerTools.AdvancedThrusters

IdlePlanetMinerTools.Alchemy

IdlePlanetMinerTools.Alloy

IdlePlanetMinerTools.AnomalyResolution

IdlePlanetMinerTools.AsteroidAutoMiner

IdlePlanetMinerTools.AsteroidHarvester

IdlePlanetMinerTools.AsteroidMiner

IdlePlanetMinerTools.AsteroidRefinedDrilling

IdlePlanetMinerTools.AsteroidScanner

IdlePlanetMinerTools.Beacon

IdlePlanetMinerTools.BottleneckOptimizations

IdlePlanetMinerTools.CargoLogistics

IdlePlanetMinerTools.Colonization

IdlePlanetMinerTools.ColonizationAdvancedEfficiency

IdlePlanetMinerTools.ColonizationAdvancedScouting

IdlePlanetMinerTools.ColonizationEfficiency

IdlePlanetMinerTools.ColonizationScouting

IdlePlanetMinerTools.ColonizationSuperiorEfficiency

IdlePlanetMinerTools.ColonizationSuperiorScouting

IdlePlanetMinerTools.ColonyAdvancedTaxIncentives

IdlePlanetMinerTools.ColonyRenegotiation

IdlePlanetMinerTools.ColonySuperiorTaxIncentives

IdlePlanetMinerTools.ColonyTaxIncentives

IdlePlanetMinerTools.ContractManager

IdlePlanetMinerTools.Craft

IdlePlanetMinerTools.Crafted

IdlePlanetMinerTools.Crafter

IdlePlanetMinerTools.CraftingEfficiency

IdlePlanetMinerTools.CraftingSpecialist

IdlePlanetMinerTools.DebrisHarvester

IdlePlanetMinerTools.DebrisScanner

IdlePlanetMinerTools.EchoFiltering

IdlePlanetMinerTools.Exploit

IdlePlanetMinerTools.FurnaceOverdrive

IdlePlanetMinerTools.FurnaceOverdrive2

IdlePlanetMinerTools.FurnaceOverdrive3

IdlePlanetMinerTools.FurnaceOverdriveEfficiency

IdlePlanetMinerTools.GameState

IdlePlanetMinerTools.HarmonicCalibration

IdlePlanetMinerTools.InsideTrader

IdlePlanetMinerTools.Inventory

IdlePlanetMinerTools.Leader

IdlePlanetMinerTools.Management

IdlePlanetMinerTools.ManagerTraining

IdlePlanetMinerTools.MarketAccelerator

IdlePlanetMinerTools.MarketInsight

IdlePlanetMinerTools.MarketManipulation

IdlePlanetMinerTools.Mine

IdlePlanetMinerTools.Modifier

IdlePlanetMinerTools.NoAdsMineBoost

IdlePlanetMinerTools.Ore

IdlePlanetMinerTools.OreTargeting

IdlePlanetMinerTools.OreYield

IdlePlanetMinerTools.Planet

IdlePlanetMinerTools.PreferredVendor

IdlePlanetMinerTools.Process

IdlePlanetMinerTools.Project

IdlePlanetMinerTools.ReactiveSweep

IdlePlanetMinerTools.Recipie

IdlePlanetMinerTools.Research

IdlePlanetMinerTools.ResourceDetails

IdlePlanetMinerTools.Room

IdlePlanetMinerTools.Rover

IdlePlanetMinerTools.RoverAdvancedLogistics

IdlePlanetMinerTools.RoverResupply

IdlePlanetMinerTools.RoverScanningModule

IdlePlanetMinerTools.SeniorLeader

IdlePlanetMinerTools.SignalStratification

IdlePlanetMinerTools.Smelt

IdlePlanetMinerTools.Smelter

IdlePlanetMinerTools.SmeltingEfficiency

IdlePlanetMinerTools.SmeltingSpecialist

IdlePlanetMinerTools.SpecialistUniversity

IdlePlanetMinerTools.SuperiorAlchemy

IdlePlanetMinerTools.SuperiorAlloyValue

IdlePlanetMinerTools.SuperiorAnomalyResolution

IdlePlanetMinerTools.SuperiorAsteroidHarvester

IdlePlanetMinerTools.SuperiorCargoHandling

IdlePlanetMinerTools.SuperiorCrafting

IdlePlanetMinerTools.SuperiorDebrisHarvester

IdlePlanetMinerTools.SuperiorEchoFiltering

IdlePlanetMinerTools.SuperiorFurnace

IdlePlanetMinerTools.SuperiorHarmonicCalibration

IdlePlanetMinerTools.SuperiorItemValue

IdlePlanetMinerTools.SuperiorLeader

IdlePlanetMinerTools.SuperiorManagerTraining

IdlePlanetMinerTools.SuperiorMining

IdlePlanetMinerTools.SuperiorReactiveSweep

IdlePlanetMinerTools.SuperiorSignalStratification

IdlePlanetMinerTools.SuperiorSpecialistUniversity

IdlePlanetMinerTools.SuperiorSurgeMapping

IdlePlanetMinerTools.SuperiorThrusters

IdlePlanetMinerTools.SurgeMapping

IdlePlanetMinerTools.Telescope1

IdlePlanetMinerTools.Telescope10

IdlePlanetMinerTools.Telescope11

IdlePlanetMinerTools.Telescope12

IdlePlanetMinerTools.Telescope13

IdlePlanetMinerTools.Telescope14

IdlePlanetMinerTools.Telescope15

IdlePlanetMinerTools.Telescope16

IdlePlanetMinerTools.Telescope17

IdlePlanetMinerTools.Telescope18

IdlePlanetMinerTools.Telescope19

IdlePlanetMinerTools.Telescope2

IdlePlanetMinerTools.Telescope20

IdlePlanetMinerTools.Telescope21

IdlePlanetMinerTools.Telescope22

IdlePlanetMinerTools.Telescope23

IdlePlanetMinerTools.Telescope24

IdlePlanetMinerTools.Telescope3

IdlePlanetMinerTools.Telescope4

IdlePlanetMinerTools.Telescope5

IdlePlanetMinerTools.Telescope6

IdlePlanetMinerTools.Telescope7

IdlePlanetMinerTools.Telescope8

IdlePlanetMinerTools.Telescope9

IdlePlanetMinerTools.Terraforming

IdlePlanetMinerTools.Thing

IdlePlanetMinerTools.add_modifier!

IdlePlanetMinerTools.add_planet!

IdlePlanetMinerTools.add_researched_project!

IdlePlanetMinerTools.add_researched_project!

IdlePlanetMinerTools.all_things

IdlePlanetMinerTools.base_effect

IdlePlanetMinerTools.base_selling_price

IdlePlanetMinerTools.crafting_plan

IdlePlanetMinerTools.delta

IdlePlanetMinerTools.doctext

IdlePlanetMinerTools.max_level

IdlePlanetMinerTools.next_projects

IdlePlanetMinerTools.ordinal

IdlePlanetMinerTools.per_level_effect

IdlePlanetMinerTools.precursor

IdlePlanetMinerTools.precursor

IdlePlanetMinerTools.precursor

IdlePlanetMinerTools.precursors

IdlePlanetMinerTools.prerequisites

IdlePlanetMinerTools.process_ingredient_scalar

IdlePlanetMinerTools.process_speed_scalar

IdlePlanetMinerTools.room_effect_factor

IdlePlanetMinerTools.to_make

Definitions

IdlePlanetMinerTools.DEFAULT_MODIFIERS — Constant

DEFAULT_MODIFIERS is the list of Modifiers that is used for operations like deltawhere none are specified.

IdlePlanetMinerTools.AdvancedAlchemy — Type

Convert an ore on a planet to an ore worth 2 levels higher

IdlePlanetMinerTools.AdvancedAlloyValue — Type

x1.2 to all alloy values

IdlePlanetMinerTools.AdvancedAnomalyResolution — Type

Detected Surges on planets within Telescopes 16 to 19 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedAsteroidHarvester — Type

All asteroids value x1.3 (multiplicative with Asteroid Harvester)

IdlePlanetMinerTools.AdvancedAsteroidScanner — Type

Frequency of asteroids increased by 30% (60% total timer reduction to spawn)

IdlePlanetMinerTools.AdvancedCargoHandling — Type

All ship cargo x 1.25

IdlePlanetMinerTools.AdvancedContractManager — Type

Increases manager slots by 1

IdlePlanetMinerTools.AdvancedCrafter — Type

All crafters speed x 1.20

IdlePlanetMinerTools.AdvancedDebrisHarvester — Type

All debris value x1.3 (multiplicative with Debris Harvester)

IdlePlanetMinerTools.AdvancedEchoFiltering — Type

Detected Surges on planets within Telescopes 4 to 7 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedFurnace — Type

All smelters speed x 1.20

IdlePlanetMinerTools.AdvancedFurnaceOverdrive — Type

Increases Furnace Overdrive smelt speed to x4

IdlePlanetMinerTools.AdvancedHarmonicCalibration — Type

Detected Surges on planets within Telescopes 8 to 11 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedItemValue — Type

x1.20 to all item values

IdlePlanetMinerTools.AdvancedLeader — Type

All leader bonuses x3

IdlePlanetMinerTools.AdvancedManagerTraining — Type

All manager benefits x1.10

IdlePlanetMinerTools.AdvancedMarketAccelerator — Type

Ability to replace the current market with the following one a second time every 8 hours

IdlePlanetMinerTools.AdvancedMarketManipulation — Type

Chance of x5 positive market

IdlePlanetMinerTools.AdvancedMining — Type

All planet mine rates x 1.25

IdlePlanetMinerTools.AdvancedOreTargeting — Type

Prioritized ore on each planet receives +30% yield bonus

IdlePlanetMinerTools.AdvancedReactiveSweep — Type

Detected Surges on planets within Telescopes 12 to 15 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedRoverResupply — Type

Planets being scanned by a rover receive x5 bonus instead of x2.5

IdlePlanetMinerTools.AdvancedSignalStratification — Type

Detected Surges on planets within Telescopes 20 to 23 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedSpecialistUniversity — Type

All manager secondary benefits x1.05

IdlePlanetMinerTools.AdvancedSurgeMapping — Type

Detected Surges on planets within Telescopes 0 to 3 are now class 3. Higher classes amplify surge effects.

IdlePlanetMinerTools.AdvancedTerraforming — Type

Increases one purchased planet's colonization level by 3 instantly

IdlePlanetMinerTools.AdvancedThrusters — Type

All ship speeds x 1.25

IdlePlanetMinerTools.Alchemy — Type

Convert an ore on a planet to an ore worth 1 level higher

IdlePlanetMinerTools.Alloy — Type

Alloy is the abstract supertype of anything that is smelted from Ore or other Alloys.

IdlePlanetMinerTools.AnomalyResolution — Type

Detected Surges on planets within Telescopes 16 to 19 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.AsteroidAutoMiner — Type

Automatically mine asteroids, even while idle

IdlePlanetMinerTools.AsteroidHarvester — Type

All asteroids value x1.3

IdlePlanetMinerTools.AsteroidMiner — Type

Ability to mine asteroids for ore

IdlePlanetMinerTools.AsteroidRefinedDrilling — Type

10% chance of an asteroid containing 20x normal value worth of alloy

IdlePlanetMinerTools.AsteroidScanner — Type

Frequency of asteroids increased by 30%

IdlePlanetMinerTools.Beacon — Type

Constructs a beacon that provides bonuses to clusters of planets

IdlePlanetMinerTools.BottleneckOptimizations — Type

Adds red/green lights to planet panel; red on shipping/cargo means planet counts as backlogging

IdlePlanetMinerTools.CargoLogistics — Type

Ability to prioritize which ore to pick up first, even if it isn't the most valuable

IdlePlanetMinerTools.Colonization — Type

Complete planet requests to colonize them and earn bonuses

IdlePlanetMinerTools.ColonizationAdvancedEfficiency — Type

Colonization costs 20% less resources

IdlePlanetMinerTools.ColonizationAdvancedScouting — Type

Increases max colonization tasks available from 4 to 5

IdlePlanetMinerTools.ColonizationEfficiency — Type

Colonization costs 20% less resources

IdlePlanetMinerTools.ColonizationScouting — Type

Increases max colonization tasks available from 3 to 4

IdlePlanetMinerTools.ColonizationSuperiorEfficiency — Type

Colonization costs 20% less resources

IdlePlanetMinerTools.ColonizationSuperiorScouting — Type

Increases max colonization tasks available from 5 to 6

IdlePlanetMinerTools.ColonyAdvancedTaxIncentives — Type

Planet upgrades reduced by 5% more per colony level (mine, speed, cargo)

IdlePlanetMinerTools.ColonyRenegotiation — Type

Allows you to renegotiate colonization requests of planets

IdlePlanetMinerTools.ColonySuperiorTaxIncentives — Type

Planet upgrades reduced by 5% more per colony level (mine, speed, cargo)

IdlePlanetMinerTools.ColonyTaxIncentives — Type

Planet upgrades reduced by 5% per colony level (mine, speed, cargo)

IdlePlanetMinerTools.ContractManager — Type

Increases manager slots by 1

IdlePlanetMinerTools.Craft — Type

Craft is the subtype of Pricess that produces Crafted items.

IdlePlanetMinerTools.Crafted — Type

Crafted is the abstract suppertype of anything that is crafted from other Things.

IdlePlanetMinerTools.Crafter — Type

Crafts items from raw materials

IdlePlanetMinerTools.CraftingEfficiency — Type

Decreases ingredients required by 20% for all crafters

IdlePlanetMinerTools.CraftingSpecialist — Type

10% chance to craft double of an item

IdlePlanetMinerTools.DebrisHarvester — Type

All debris value x1.3

IdlePlanetMinerTools.DebrisScanner — Type

Allows discovery of debris fields that provide valuable items

IdlePlanetMinerTools.EchoFiltering — Type

Detected Surges on planets within Telescopes 4 to 7 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.Exploit — Type

Exploit is the process by which a new planet is discovered.

IdlePlanetMinerTools.FurnaceOverdrive — Type

Ability to put one furnace into overdrive. Smelt speed x2, but ingredients cost +50%

IdlePlanetMinerTools.FurnaceOverdrive2 — Type

Ability to put a second furnace into overdrive

IdlePlanetMinerTools.FurnaceOverdrive3 — Type

Ability to put a third furnace into overdrive

IdlePlanetMinerTools.FurnaceOverdriveEfficiency — Type

Decreases Furnace Overdrive extra ingredient cost to +20%

IdlePlanetMinerTools.GameState — Type

GameState()

GameState encapsulates the state of the IdlePlanetMiner game.

IdlePlanetMinerTools.HarmonicCalibration — Type

Detected Surges on planets within Telescopes 8 to 11 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.InsideTrader — Type

+15% chance of positive market fluctuations on future changes

IdlePlanetMinerTools.Inventory — Type

An Inventory represents a collection of Things.

IdlePlanetMinerTools.Leader — Type

Ability to elect a leader among the managers. Leader bonuses x2

IdlePlanetMinerTools.Management — Type

Hire and assign managers to planets

IdlePlanetMinerTools.ManagerTraining — Type

All manager benefits x1.10

IdlePlanetMinerTools.MarketAccelerator — Type

Ability to replace the current market with the following one, once every 8 hours

IdlePlanetMinerTools.MarketInsight — Type

Ability for forecast upcoming market changes

IdlePlanetMinerTools.MarketManipulation — Type

Chance of x4 positive market

IdlePlanetMinerTools.Mine — Type

Mine is the subtype of Process that generates Ore.

IdlePlanetMinerTools.Modifier — Type

Modifier is the abstract supertype for anything that can modify various arithmetic factors in the game, for example Projects or Rooms that modify smelting ingredients.

IdlePlanetMinerTools.NoAdsMineBoost — Type

NoAdsMineBoost is the modifier that provides a 1.2 increase in mining rate when you pay for no ads.

IdlePlanetMinerTools.Ore — Type

Ore is the abstract supertype for anything that is mined.

IdlePlanetMinerTools.OreTargeting — Type

Prioritized ore on each planet receives +15% yield bonus

IdlePlanetMinerTools.OreYield — Type

OreYield represents one of a Planet's ore yields.

IdlePlanetMinerTools.Planet — Type

Planet represents a Planet.

IdlePlanetMinerTools.PreferredVendor — Type

Price of building new smelters and crafters reduced by 50%

IdlePlanetMinerTools.Process — Type

Process is the abstract supertype for transformative operations.

IdlePlanetMinerTools.Project — Type

Project represents an Idle Planet Miner project.

IdlePlanetMinerTools.ReactiveSweep — Type

Detected Surges on planets within Telescopes 12 to 15 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.Recipie — Type

Recipie describes how to make Alloys, crafted items, projects, etc, anything that required some number of Things.

IdlePlanetMinerTools.Research — Type

Research is the subtype of Process that produces Projects.

IdlePlanetMinerTools.ResourceDetails — Type

Adds a detail panel with resource stats. Accessed from resource panel

IdlePlanetMinerTools.Room — Type

Room is the abstract supertype for all space ship rooms. Each room has a level field.

Associated with each Room type is a base_effect and a per_level_effect, and a max_level.

IdlePlanetMinerTools.Rover — Type

Ability to scan planets for valuable items

IdlePlanetMinerTools.RoverAdvancedLogistics — Type

Unlocks an additional mission choice for new scans

IdlePlanetMinerTools.RoverResupply — Type

Planets being scanned by a rover receive x2.5 bonus to mining speed, ship speed, and cargo

IdlePlanetMinerTools.RoverScanningModule — Type

+5% chance of rovers finding all rewards

IdlePlanetMinerTools.SeniorLeader — Type

Ability to elect a second Leader among 5+ star managers

IdlePlanetMinerTools.SignalStratification — Type

Detected Surges on planets within Telescopes 20 to 23 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.Smelt — Type

Smelt is the subtype of Pricess that produces Alloys.

IdlePlanetMinerTools.Smelter — Type

Refine ores into more valuable materials

IdlePlanetMinerTools.SmeltingEfficiency — Type

Decreases ingredients required by 20% for all smelters

IdlePlanetMinerTools.SmeltingSpecialist — Type

10% chance to smelt double of an alloy

IdlePlanetMinerTools.SpecialistUniversity — Type

All manager secondary benefits x1.05

IdlePlanetMinerTools.SuperiorAlchemy — Type

Convert an ore on a planet to an ore worth 3 levels higher

IdlePlanetMinerTools.SuperiorAlloyValue — Type

x1.2 to all alloy values

IdlePlanetMinerTools.SuperiorAnomalyResolution — Type

Detected Surges on planets within Telescopes 16 to 19 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorAsteroidHarvester — Type

All asteroids value x1.3 (multiplicative with Advanced Asteroid Harvester)

IdlePlanetMinerTools.SuperiorCargoHandling — Type

All ship cargo x 1.25

IdlePlanetMinerTools.SuperiorCrafting — Type

All crafters speed x1.20

IdlePlanetMinerTools.SuperiorDebrisHarvester — Type

All debris value x1.3 (multiplicative with previous upgrades)

IdlePlanetMinerTools.SuperiorEchoFiltering — Type

Detected Surges on planets within Telescopes 4 to 7 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorFurnace — Type

All smelters speed x1.20

IdlePlanetMinerTools.SuperiorHarmonicCalibration — Type

Detected Surges on planets within Telescopes 8 to 11 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorItemValue — Type

x1.20 to all item values

IdlePlanetMinerTools.SuperiorLeader — Type

All leader bonuses x4

IdlePlanetMinerTools.SuperiorManagerTraining — Type

All manager benefits x1.10

IdlePlanetMinerTools.SuperiorMining — Type

All planet mine rates x 1.25

IdlePlanetMinerTools.SuperiorReactiveSweep — Type

Detected Surges on planets within Telescopes 12 to 15 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorSignalStratification — Type

Detected Surges on planets within Telescopes 20 to 23 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorSpecialistUniversity — Type

All manager secondary benefits x1.05

IdlePlanetMinerTools.SuperiorSurgeMapping — Type

Detected Surges on planets within Telescopes 0 to 3 are now class 4. Higher classes amplify surge effects.

IdlePlanetMinerTools.SuperiorThrusters — Type

All ship speeds x 1.25

IdlePlanetMinerTools.SurgeMapping — Type

Detected Surges on planets within Telescopes 0 to 3 are now class 2. Higher classes amplify surge effects.

IdlePlanetMinerTools.Telescope1 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope10 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope11 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope12 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope13 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope14 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope15 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope16 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope17 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope18 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope19 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope2 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope20 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope21 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope22 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope23 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope24 — Type

Extends vision to planets 74 - 76

IdlePlanetMinerTools.Telescope3 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope4 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope5 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope6 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope7 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope8 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Telescope9 — Type

Extends vision to 3 more planets

IdlePlanetMinerTools.Terraforming — Type

Increases one purchased planet's colonization level by 2 instantly

IdlePlanetMinerTools.Thing — Type

Thing is the abstract supertype of Ore, Alloy and Crafted.

Every subtype of Thing has a count field.

IdlePlanetMinerTools.add_modifier! — Method

add_modifier!(m::Modifier, gs::GameState)

Adds the specified modifier to the GameState. If the same type of modifier is already present, it is replaced by the new one.

Returns the GameState.

IdlePlanetMinerTools.add_planet! — Method

add_planet!(planet::Planet, gs::GameState)

Adds the specified Planet to the GameState if it is not already present.

Returns the GameState.

IdlePlanetMinerTools.add_researched_project! — Function

add_researched_project!(add_this::Type{<:Project}, to::Vector{Modifier}=DEFAULT_MODIFIERS)

Add add_this and the transitive closure of its prerequisites to to if they are not already present.

IdlePlanetMinerTools.add_researched_project! — Method

add_researched_project!(add_this::Type{<:Project}, gs::GameState)

Add add_this to the GqameState if its prerequisites are met.

Returns the GameStqate.

IdlePlanetMinerTools.all_things — Method

all_things()

returns a list of all concrete subtypes of Thing in canonical order.

IdlePlanetMinerTools.base_effect — Function

base_effect(::Type{<:Room})

Returns the modifier factor for the room type at level 1.

IdlePlanetMinerTools.base_selling_price — Method

base_selling_price(::Type{<:Thing})
base_selling_price(::Thing)

Returns the base selling price for the thing.

For instances it multiplies the unit price by the count.

IdlePlanetMinerTools.crafting_plan — Function

crafting_plan(inv::Inventory, modifiers = DEFAULT_MODIFIERS)

Appliies Recipies to the ~Inventory` until there are no negative item counts or there is no known recipie to resolve the current shortage.

Returns a list of CraftingActions and the final Inventory.

IdlePlanetMinerTools.delta — Function

delta(r::Recipie, modifiers = DEFAULT_MODIFIERS)

Returns an Inventory that would be the effect of applying the Recipie. modifiers is a vector of the player's current Modifiers.

It appears that the game rounds the result to the nearest integer. I've not yet decided if that should happen here.

IdlePlanetMinerTools.doctext — Method

doctext(::Type{<:Modifier})

Returns a text string defining the Modifier.

IdlePlanetMinerTools.max_level — Function

max_level(::Type{<:Room})

Returns the maximum allowed value for the level field for the specified room type.

IdlePlanetMinerTools.next_projects — Function

next_projects(modifiers=DEFAULT_MODIFIERS)

Assuming that the projects in modifiers have already been researches, what are the next projects to be researches?

IdlePlanetMinerTools.ordinal — Method

ordinal(thing)

Returns the ordinal number for thing, where thing is either a subtype of Thing or an instance of a subtype of Thing.

IdlePlanetMinerTools.per_level_effect — Function

per_level_effect(::Type{<:Room})

Returns the modifier factor for the room type for each level above level 1.

IdlePlanetMinerTools.precursor — Method

precursor(::Planet)

The precursor of a Planet is a Telescope.

IdlePlanetMinerTools.precursor — Method

function precursor(o::Type{<:Ore})

The precursor of an Ore is any of the Planets that provide that Ore.

We restrict the result to those planets with the lowest Telescope number.

IdlePlanetMinerTools.precursor — Method

precursor(t::Type{<:Thing})

The precursor of any subtype of Thing that is not a subtype of Ore is AllOf the ingredients of its Recipie.

IdlePlanetMinerTools.precursors — Function

precursor(x)

Returns the precursor (for planning purposes) of x.

IdlePlanetMinerTools.prerequisites — Method

prerequisites(::Type{<:Project})

Returns a vector of the immediate prerequisites that are required before researching the specified Project.

IdlePlanetMinerTools.process_ingredient_scalar — Method

process_ingredient_scalar(::Process, ::Modifier)

Returns the scalar multiplier for the amount of ingredients required for the Process as adjusted by Modifier.

IdlePlanetMinerTools.process_speed_scalar — Method

process_speed_scalar(::Process, ::Type{<:Modifier})

Returns the scalar multiplier for the speed of the Process as adjusted by Modifier.

IdlePlanetMinerTools.room_effect_factor — Method

room_effect_factor(::Room)

Computes the effect factor for the Room based on its type and level. This says nothing about what is modified. It just represents a calculation that tis common to all rooms.

IdlePlanetMinerTools.to_make — Method

to_make(type)

Returns an instance of the subtype of Process that can make an instance of type.