IdlePlanetMinerTools

Documentation for IdlePlanetMinerTools.

using IdlePlanetMinerTools

need = 7 * rx"gravity chamber".delta

have = GravityChamber(4) + AdvancedComputer(42) + BasicComputer(64) + Circuit(902)

actions, result = crafting_plan(have - need)

(CraftingAction[CraftingAction(Recipie(GravityChamber, Inventory(Thing[AdvancedComputer(-60), NuclearReactor(-1), GravityChamber(1)]), 24300), 3)], Inventory(Thing[Circuit(902), BasicComputer(64), AdvancedComputer(282), NuclearReactor(4), GravityChamber(0)]))

IdlePlanetMinerTools.crafting_plan

IdlePlanetMinerTools.ordinal

IdlePlanetMinerTools.crafting_plan — Method

crafting_plan(inv::Inventory)

Appliies Recipies to the ~Inventory` until there are no negative item counts or there is no known recipie to resolve the current shortage.

Returns a list of CraftingActions and the final Inventory.

IdlePlanetMinerTools.ordinal — Method

ordinal(thing)

Returns the ordinal number for thing, where thing is either a subtype of Thing or an instance of a subtype of Thing.