IdlePlanetMinerTools

Documentation for IdlePlanetMinerTools.

Things and Inventory

Mining, smelting and crafting produce instances of subtypes of Ore, Alloy, or Crafted, respectively. These are all subtypes of Thing. Each concrete subtype of Thing has a count field indicating a quantity of that Thing.

One can make a Thing using a concrete constructor

using IdlePlanetMinerTools

Battery(3)

Battery(3)

or using the @t_str macro, which makes the named thing with a count of 1.

t"battery"

Battery(1)

Note that for convenience, @t_str uses a dictionary of all concrate Thing subtypes and finds the closest string match.

An Inventory represents a collection of disperate Things.

For convenience, there is a total ordering defined on all concrete subtypes of Thing.

@@ Arithmetic

Arithmetic operations are defined for Things and Inventorys:

using IdlePlanetMinerTools

need = 7 * delta(rx"gravity chamber")

have = GravityChamber(4) + AdvancedComputer(42) + BasicComputer(64) + Circuit(902)

actions, result = crafting_plan(have - need)

(CraftingAction[CraftingAction(Recipie(GravityChamber, Inventory(Thing[AdvancedComputer(-60), NuclearReactor(-1)]), 24300), 3)], Inventory(Thing[Circuit(902), BasicComputer(64), AdvancedComputer(282), NuclearReactor(4), GravityChamber(0)]))

IdlePlanetMinerTools.Alloy

IdlePlanetMinerTools.Crafted

IdlePlanetMinerTools.Ore

IdlePlanetMinerTools.Thing

IdlePlanetMinerTools.all_things

IdlePlanetMinerTools.crafting_plan

IdlePlanetMinerTools.ordinal

IdlePlanetMinerTools.Alloy — Type

Alloy is the abstract supertype of anything that is smelted from Ore or other Alloys.

IdlePlanetMinerTools.Crafted — Type

Crafted is the abstract suppertype of anything that is crafted from other Things.

IdlePlanetMinerTools.Ore — Type

Ore is the abstract supertype for anything that is mined.

IdlePlanetMinerTools.Thing — Type

Thing is the abstract supertype of Ore, Alloy and Crafted.

IdlePlanetMinerTools.all_things — Method

all_things()

returns a list of all concrete subtypes of Thing in canonical order.

IdlePlanetMinerTools.crafting_plan — Method

crafting_plan(inv::Inventory)

Appliies Recipies to the ~Inventory` until there are no negative item counts or there is no known recipie to resolve the current shortage.

Returns a list of CraftingActions and the final Inventory.

IdlePlanetMinerTools.ordinal — Method

ordinal(thing)

Returns the ordinal number for thing, where thing is either a subtype of Thing or an instance of a subtype of Thing.